Testing build and local pip install
