+++ tests/hd131339a.py
@@ -1,10 +1,8 @@
-import sys
-sys.path.append('..')
-from backtrack.backtrack import backtrack
+from backtracks import backtrack
 
 if __name__ == '__main__':
-    mytrack = backtrack("HD 131399 A", "scorpions1b_orbitizelike.csv", nearby_window=0.5)
+    mytrack = backtrack.system("HD 131399 A", "scorpions1b_orbitizelike.csv", nearby_window=0.5)
     results = mytrack.fit()
-    mytrack.generate_plots()
+    mytrack.generate_plots(days_backward=2600, days_forward=1200)
     mytrack.save_results()
     mytrack.load_results()
